data_generator/requirements.txt
confluent-kafka==2.3.0
fastavro==1.9.4
faker==21.0.0
python-dateutil==2.8.2
pyyaml==6.0.1
structlog==24.1.0
prometheus-client==0.19.0
numpy==1.26.3
scipy==1.12.0
